ai_assistant/requirements.txt
langchain
langchain-community
transformers
groq
fastapi
uvicorn
pymongo
python-dotenv
langchain_groq